:hammer:Reorganiza capretas, componetes y reorganiza codigo en pagina profile y routesApp.jsx
