+++ src/routes/RoutesApp.jsx
@@ -5,26 +5,6 @@ import ProfilePage from "../pages/profile/ProfilePage.jsx";
 import NotificationPage from "../pages/NotificationPage.jsx";
 import SignupPage from "../pages/SignupPage.jsx";
 import ForgotPasswordPage from "../pages/ForgotPasswordPage.jsx";
-import SidebarComponent from "../components/Sidebar.jsx";
-
-
-function Perfil() {
-	return (
-	  <div className=" bg-primary-50">
-		<SidebarComponent/>
-		<ProfilePage />
-	  </div>
-	);
-  }
-
-  function Notificaciones() {
-	return (
-	  <div>
-		<SidebarComponent/>
-		<NotificationPage />
-	  </div>
-	);
-  }
 
 const RoutesApp = () => {
 	return (
@@ -32,7 +12,7 @@ const RoutesApp = () => {
 			<Route path="*" element={<NotFoundPage/>}></Route>
 			<Route path="/" element={<Navigate to="/login" />}></Route>
 			<Route path="/login" element={<LoginPage />}></Route>
-			<Route path="/profile" element={<Perfil/>} />
+			<Route path="/profile" element={<ProfilePage/>} />
 			<Route path="/signup" element={<SignupPage/>} />
 			<Route path="/notification" element={<NotificationPage/>} />
 			<Route path="/forgot-password" element={<ForgotPasswordPage/>} />
